+++ Events/SuitSoldEvent.cs
@@ -7,15 +7,14 @@ using Utilities;
 namespace EddiEvents
 {
     [PublicAPI]
-    public class SuitPurchasedEvent ( DateTime timestamp, Suit suit, int? price, List<SuitMod> mods ) 
+    public class SuitSoldEvent ( DateTime timestamp, Suit suit, int? price, List<SuitMod> mods ) 
         : Event( timestamp, NAME )
     {
-        public const string NAME = "Suit purchased";
-        public const string DESCRIPTION = "Triggered when you buy a space suit";
+        public const string NAME = "Suit sold";
+        public const string DESCRIPTION = "Triggered when you sell a space suit";
         public static readonly string[] SAMPLES = {
-            @"{ ""timestamp"":""2021-05-30T00:55:52Z"", ""event"":""BuySuit"", ""Name"":""TacticalSuit_Class3"", ""Name_Localised"":""$TacticalSuit_Class1_Name;"", ""Price"":18750000, ""SuitID"":1701142753535138, ""SuitMods"":[ ""suit_adsmovementspeed"", ""suit_increasedo2capacity"" ] }",
-            @"{ ""timestamp"":""2021-06-04T07:54:42Z"", ""event"":""BuySuit"", ""Name"":""UtilitySuit_Class3"", ""Name_Localised"":""$UtilitySuit_Class1_Name;"", ""Price"":11250000, ""SuitID"":1701622088710220, ""SuitMods"":[ ""suit_improvedradar"" ] }",
-            @"{ ""timestamp"":""2021-06-05T04:29:13Z"", ""event"":""BuySuit"", ""Name"":""ExplorationSuit_Class3"", ""Name_Localised"":""$ExplorationSuit_Class1_Name;"", ""Price"":11250000, ""SuitID"":1701699758193117, ""SuitMods"":[ ""suit_improvedarmourrating"" ] }"
+            @"{ ""timestamp"":""2025-02-03T08:59:16Z"", ""event"":""SellSuit"", ""SuitID"":1700534679057635, ""SuitMods"":[ ""suit_increasedshieldregen"" ], ""Name"":""tacticalsuit_class3"", ""Name_Localised"":""$TacticalSuit_Class1_Name;"", ""Price"":825000 }",
+            @"{ ""timestamp"":""2021-09-05T04:56:23Z"", ""event"":""SellSuit"", ""SuitID"":1701699758193117, ""SuitMods"":[ ""suit_improvedarmourrating"" ], ""Name"":""explorationsuit_class3"", ""Name_Localised"":""$ExplorationSuit_Class1_Name;"", ""Price"":825000 }"
         };
 
         [PublicAPI("The space suit, as an object")]
@@ -24,7 +23,7 @@ namespace EddiEvents
         [PublicAPI( @"The space suit's grade" )]
         public int grade { get; } = suit?.grade ?? 1;
 
-        [PublicAPI("The price paid for the space suit")]
+        [PublicAPI("The space suit's sell price")]
         public int? price { get; } = price;
 
         [PublicAPI( @"The suit's modifications (as objects)" )]
@@ -48,7 +47,7 @@ namespace EddiEvents
                 }
             }
 
-            events.Add( new SuitPurchasedEvent( timestamp, suit, price, mods ) { raw = line, fromLoad = fromLogLoad } );
+            events.Add( new SuitSoldEvent( timestamp, suit, price, mods ) { raw = line, fromLoad = fromLogLoad } );
             return true;
         }
     }