+++ SpeechService/SpeechConversions/PhoneticShipManufacturer.cs
@@ -6,7 +6,7 @@ namespace EddiSpeechService.SpeechConversions
     {
         public static string getPhoneticShipManufacturer(string val2)
         {
-            var phoneticManufacturer = ShipManufacturer.SpokenManufacturer(val2);
+            var phoneticManufacturer = Manufacturer.SpokenManufacturer(val2);
             if (!string.IsNullOrEmpty( phoneticManufacturer ) )
             {
                 return phoneticManufacturer.Trim();